refactor(entity-account): parse ids via a union schema (LIVE-36924)
Review pass on #21554.

`AnyAccountIdSchema` is now a `z.union` and the two parsers delegate to it,
instead of discriminating on `+` by hand. The members are mutually exclusive by
construction, so the union never disambiguates and its order is irrelevant —
covered by a test.

`encodeTokenAccountId` drops its manual guard: `TokenAccountIdSchema` already
rejects an empty part after the `+` and a second `+`. The two tests that
covered the guard now prove the schema does the rejecting.

`getParentId` casts instead of parsing: `TokenAccountId`'s invariant is exactly
`AccountId`'s for the part before the `+`, so the parse re-validated what the
argument's type guarantees.

Files split as the reviewer asked, following `entity-currency-fiat`: schemas and
their inferred types in schema.ts, functions in utils.ts. No types.ts — every
type here is schema-inferred, and that is what schema.ts is for.
