+++ domain/entity/account/src/index.ts
@@ -1,1 +1,2 @@
-export * from "./account-id";
+export * from "./schema";
+export * from "./utils";
